controller usuario

+++ ControllerJs/usuarioController.js
@@ -1,3 +1,44 @@
+function limparForm()
+{
+    var fdados = document.getElementById("fusuario");
+    fdados.nome.value="";
+    fdados.email.value="";
+    fdados.senha.value="";
+    fdados.telefone.value="";
+    fdados.cpf.value="";
+    fdados.privilegio.value="";
+    fdados.sexo.value="M";
+    fdados.cep.value="";
+    fdados.rua.value="";
+    fdados.bairro.value="";
+    fdados.numero.value="";
+}
+
+function validarCampos()
+{
+  const nome = document.getElementById("nome").value;
+  const email = document.getElementById("email").value;
+  const senha = document.getElementById("senha").value;
+  const telefone = document.getElementById("telefone").value;
+  const cpf = document.getElementById("cpf").value;
+  const privilegio = document.getElementById("privilegio").value;
+  const sexo = document.getElementById("sexo").value;
+  const cep = document.getElementById("cep").value;
+  const rua = document.getElementById("rua").value;
+  const bairro = document.getElementById("bairro").value;
+  const numero = document.getElementById("numero").value;
+  
+  if (nome != "" && email != "" && senha != "" && telefone != "" && cpf != "" && privilegio != "" && sexo != "" && cep != "" && rua != "" && bairro != "" && numero != "")
+  {
+    cadUsuario();
+  }
+  else
+  {
+    alert("Campo(s) Não Preenchido(s)")
+  }
+  limparForm();
+}
+
 function buscarUsuario() {
     let filtro = document.getElementById("filtro").value
     const resultado = document.getElementById("resultado");
@@ -81,23 +122,36 @@ function buscarUsuario() {
     }
 }
 
-function excluirUsuario(id){
-    const URL = "http://localhost:8080/apis/usuario/excluir/" + id;
+function editarUsuario(id) {
     
-    fetch(URL, {
-        headers: {
-            'Accept': 'application/json',
-            'Content-Type': 'application/json'
-        },
-        method: 'DELETE'
-    })
-        .then((response) => {
-            return response.json();
+    window.location.href = "../TelasCadastros/cadUsuario.html?id="+id;
+}
+
+function excluirUsuario(id) 
+{
+
+    const confirmacao = confirm("Tem certeza que deseja excluir este usuario ?");
+    if (confirmacao)
+    {
+        const URL = "http://localhost:8080/apis/usuario/excluir/" + id;
+
+        fetch(URL, {
+            headers: {
+                'Accept': 'application/json',
+                'Content-Type': 'application/json'
+            },
+            method: 'DELETE'
         })
-        .then((json) => {
-            alert(JSON.stringify(json));
-        })
-        .catch((error) => console.error(error));
+            .then((response) => {
+                return response.json();
+            })
+            .then((json) => {
+                alert("Usuario Excluido Com Sucesso");
+                window.location.reload();
+            })
+            .catch((error) => console.error("Erro ao excluir o usuario:", error));
+    } 
+    
 }
 
 function editarUsuario(id){
@@ -123,23 +177,47 @@ function editarUsuario(id){
 }
 
 function cadUsuario() {
-    const URL = "http://localhost:8080/apis/usuario/gravar";
+    
     var fusuario = document.getElementById("fusuario");
     var jsontext = JSON.stringify(Object.fromEntries(new FormData(fusuario)));
-    
-    fetch(URL, {
-        headers: {
-            'Accept': 'application/json',
-            'Content-Type': 'application/json'
-        },
-        method: 'POST', body: jsontext
-    })
-        .then((response) => {
-            return response.json();
+    var cod = document.getElementById("codUsuario").value;
+    if(cod) // existe, atualiza
+    {
+        const URL = "http://localhost:8080/apis/usuario/atualizar"
+        fetch(URL, {
+            headers: {
+                'Accept': 'application/json',
+                'Content-Type': 'application/json'
+            },
+            method: 'PUT', body: jsontext
         })
-        .then((json) => {
-            alert(JSON.stringify(json));
-            fusuario.reset();
+            .then((response) => {
+                return response.json();
+            })
+            .then((json) => {
+                alert("Usuario Alterado Com Sucesso");
+                fusuario.reset();
+            })
+            .catch((error) => console.error(error))
+
+    }
+    else
+    {
+        const URL = "http://localhost:8080/apis/usuario/gravar"
+        fetch(URL, {
+            headers: {
+                'Accept': 'application/json',
+                'Content-Type': 'application/json'
+            },
+            method: 'POST', body: jsontext
         })
-        .catch((error) => console.error(error))
+            .then((response) => {
+                return response.json();
+            })
+            .then((json) => {
+                alert("Usuario Cadastrado Com Sucesso");
+                fusuario.reset();
+            })
+            .catch((error) => console.error(error))
+    }
 }
